requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
streamlit>=1.30
plotly>=5.18
anthropic>=0.40
